fixup! [#704] Add env variable
